requirements.txt
python-dotenv==1.0.1
psycopg2==2.9.9
notebook==7.1.0
pandas==2.2.1
sqlalchemy==2.0.27